clients/python3/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dice52"
version = "0.1.0"
description = "Dice52-PQ: Post-Quantum Ratcheting Protocol"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Dice52 Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "kyber-py>=0.2.0",
    "dilithium-py>=0.1.0",
    "pycryptodome>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "build",
]

[project.scripts]
dice52-demo = "dice52.demo:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dice52*"]
